SiteMap: Draft 3 Login and Draft 2 Event account
Updated the error handling messages to the site map and ensured that all types of errors are pointed to the correct error message. Ensuring that no bugs are within the spectrum, our goal will be to send error messages to the user to notify if either something within the system went wrong or if user information entered is invalid. Also color-coded the Event account to ensure more readability within different functionalities within the application itself.