Added userid, Removed Looping Error in LogService, Fixing SQLQuery
Dictionary was not logging UserID, fixed that by having it log TEMP USER, but that might need to change. Query was also not logging timestamp but required timestamp variable, added that. Might need to change almost all to varchar, rn they're ints in the database.
CREATE was looping because kept calling SQLGenerator inside SQLGenerator, fixed that but need to fix query to ensure it's logging correctly into database.
Fixed SQL names for columns to match the actual column names.
Changed connectionString to correct user and password?